5519 - Dependencies upgrade: date-fns
