+++ package.json
@@ -75,7 +75,7 @@
     "@types/passport-google-oauth20": "^2.0.17",
     "@types/passport-jwt": "^4.0.1",
     "@types/passport-local": "^1.0.38",
-    "@types/pg": "^8.15.6",
+    "@types/pg": "^8.16.0",
     "@types/react": "^19.2.7",
     "@types/react-csv": "^1.1.2",
     "@types/react-dom": "^19.2.3",
@@ -141,7 +141,7 @@
     "d3": "^7.9.0",
     "d3-interpolate-path": "^2.3.0",
     "d3-tip": "^0.9.1",
-    "date-fns": "^2.28.0",
+    "date-fns": "^4.1.0",
     "diff": "^5.2.0",
     "diff-dom": "^5.1.4",
     "dom-serializer": "^2.0.0",
