+++ src/utils/dates/getCurrentYear.ts
@@ -1,0 +1,5 @@
+import { getYear } from 'date-fns'
+
+export const getCurrentYear = (): number => {
+  return getYear(new Date())
+}
